It is starting to take shape.
The shutdown thread now not only sets the shutdown flag,
also breaks blocking IO in the form of selects by closing the
write end of a dummy pipe. The selects see the pipe named killpipe
closing and unblock, which leads to the termination of the thread functions.

Also some bugfixes on the ExecCall processing. OpenCall processing is
still badly broken.